+++ src/mainpage.js
@@ -1,17 +1,4 @@
 import { Link, useNavigate } from 'react-router-dom';
-import axios from "axios";
-// import Blank from './blank';
-// import Components from './components';
-// import Friends from './friends';
-import Groups from './groups';
-import MarcketPlace from './marketplace';
-import Messeges from './messages';
-// import Modal from './modal';
-import NewsFeed2 from './newsfeed-2';
-import Profile from './profile';
-import Settings from './settings';
-import SignIn from './sign-in';
-import SignUp from './sign-up';
 import PostList from './postlist';
 import PostForm from './postform';
 // import Widgets from './widgets';
@@ -208,7 +195,7 @@ function MainPage() {
             className="form-control"
             id="postForm"
             rows={2}
-            placeholder="What's on your mind, Arthur?"
+            placeholder="What's on your mind, username?"
             defaultValue={""}
           />
           </div>
@@ -229,12 +216,12 @@ function MainPage() {
             className="btn btn-link post-form-btn btn-sm"
             >
             <img
-              src="assets/images/icons/theme/tag-friend.png"
+              src="assets/images/icons/theme/smile.png"
               alt="post form icon"
             />{" "}
-            <span>Tag Friends</span>
+            <span>Moods</span>
             </button>
-            <button
+            {/* <button
             type="button"
             className="btn btn-link post-form-btn btn-sm"
             >
@@ -243,7 +230,7 @@ function MainPage() {
               alt="post form icon"
             />{" "}
             <span>Check In</span>
-            </button>
+            </button> 
           </div>
           <div className="col-md-3 text-right">
             <button type="button" className="btn btn-primary btn-sm">
